+++ src/effects/builtin_collection/qml/Audacity/BuiltinEffectsCollection/BigParameterKnob.qml
@@ -97,7 +97,7 @@ Item {
             horizontalAlignment: Qt.AlignHCenter
         }
 
-        KnobControl {
+        M3Knob {
             id: knob
 
             navigation.accessible.name: root.parameter["title"]